+++ web/modules/custom/os2forms_selvbetjening/src/Helper/FormHelper.php
@@ -2,27 +2,14 @@
 
 namespace Drupal\os2forms_selvbetjening\Helper;
 
-use Drupal\Core\Config\ConfigFactoryInterface;
 use Drupal\Core\Form\FormStateInterface;
+use Drupal\Core\StringTranslation\StringTranslationTrait;
 
 /**
  * Form Helper class, for altering forms.
  */
 class FormHelper {
-
-  /**
-   * The ConfigFactoryInterface.
-   *
-   * @var \Drupal\Core\Config\ConfigFactoryInterface
-   */
-  private ConfigFactoryInterface $config;
-
-  /**
-   * Constructs a FormHelper.
-   */
-  public function __construct(ConfigFactoryInterface $config) {
-    $this->config = $config;
-  }
+  use StringTranslationTrait;
 
   /**
    * Allows altering of forms.
@@ -30,12 +17,19 @@ class FormHelper {
   public function formAlter(array &$form, FormStateInterface $form_state, string $form_id) {
     // Add description to the message body section of the email handler.
     if ('webform_handler_form' === $form_id && 'email' === ($form['#webform_handler_plugin_id'] ?? NULL)) {
+      $form['settings']['message']['body']['#description'] = $this->t('Use the default email body or define you own custom one. See <a href="https://os2forms.os2.eu/mail-tekster">OS2Forms Loop</a> for other standards and examples.');
+    }
 
-      $config = $this->config->get('os2forms_selvbetjening');
+    $webform_category_description = $this->t('Externally: Citizen. Internally: Employees');
 
-      if ($email_body_description = $config->get('email_body_description')) {
-        $form['settings']['message']['body']['#description'] = $email_body_description;
-      }
+    // Add description to category choice in Webform Settings.
+    if ('webform_settings_form' === $form_id) {
+      $form['general_settings']['category']['#description'] = $webform_category_description;
+    }
+
+    // Add description to category choice when adding new Webform.
+    if ('webform_add_form' === $form_id) {
+      $form['category']['#description'] = $webform_category_description;
     }
   }